Added automatic rerender on adding item and implemented error page
The button on the error page navigates you to the logged out homepage, so if you were logged in, you'll automatically get logged out. Don't think this is super important to have done for our presentation today though, just showing that it works. If we have time we can try to fix it, worst case scenario we can just comment out the button